src/main/webapp/app/entities/rotas-ofertas/route/rotas-ofertas-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IRotasOfertas } from '../rotas-ofertas.model';
import { RotasOfertasService } from '../service/rotas-ofertas.service';

import { RotasOfertasRoutingResolveService } from './rotas-ofertas-routing-resolve.service';

describe('RotasOfertas routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: RotasOfertasRoutingResolveService;
  let service: RotasOfertasService;
  let resultRotasOfertas: IRotasOfertas | null | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(RotasOfertasRoutingResolveService);
    service = TestBed.inject(RotasOfertasService);
    resultRotasOfertas = undefined;
  });

  describe('resolve', () => {
    it('should return IRotasOfertas returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultRotasOfertas = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultRotasOfertas).toEqual({ id: 123 });
    });

    it('should return null if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultRotasOfertas = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultRotasOfertas).toEqual(null);
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse<IRotasOfertas>({ body: null })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultRotasOfertas = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultRotasOfertas).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
